src/consumers/pyproject.toml
[project]
name = "consumer-rewrite"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
      "faust-streaming>=0.11.3",
    "aiokafka==0.10.0",
    "confluent-kafka>=2.14.0",
    "redis>=7.4.0",
    "psycopg[binary,pool]>=3.3.3",
    "dotenv>=0.9.9",

]
